+++ Fireworks/Calo/interface/FWFromSliceSelector.h
