+++ Fireworks/Core/interface/FWDetailViewCanvas.icc
