+++ Fireworks/Core/interface/FWDetailViewGL.icc
