+++ Fireworks/Core/interface/FWTEventList.h
@@ -2,7 +2,7 @@
 #define Fireworks_Core_FWTEventList_h
 
 // There was a bug in ROOT ... fixed on Dec 9 2009:
-//   http://root.cern.ch/viewcvs/trunk/tree/tree/src/TEventList.cxx?view=log
+//   http://root.cern.ch/viewcvs/trunk/tree/tree/interface/TEventList.cxx?view=log
 //
 // We need to keep this intermediate class until we switch to
 // root-5.26 or later.
